[MCP Extract] Refactor: extract JsonSchemaConfigurationSection component
Preparatory for https://github.com/dust-tt/tasks/issues/2640

Description
---
Extract the JSON schema configuration part of 'Extract' action
configuration in builder as a single component, to be reused by
Extract MCP action and any future internal action requiring a JSON
schema configuration

Risk
---
standard

Deploy
---
front
